Move Mirth message builders into submodule
